+++ lib/rubocop/cop/mixin/endpoint_helper.rb
@@ -7,7 +7,7 @@ module RuboCop
       extend NodePattern::Macros
 
       def_node_matcher :http_method_node?, <<~PATTERN
-        (block (send _ {:get :post :put :patch :delete} ...) ...)
+        (block $(send _ {:get :post :put :patch :delete} ...) ...)
       PATTERN
 
       def_node_matcher :params_node?, <<~PATTERN